requirements.txt
cchardet==2.1.7
curl_cffi==0.6.2
faust_cchardet==2.1.19
keyboard==0.13.5
lxml==4.9.0
more_itertools==10.3.0
mysql_connector_repackaged==0.3.1
pandas==2.0.3
psutil==5.9.2
pyap==0.3.1
PyGetWindow==0.0.9
PyWinCtl==0.3
requests==2.31.0
selenium==4.21.0
undetected_chromedriver==3.5.4
unicode_slugify==0.1.5
usaddress==0.5.10
